+++ packages/api/drizzle/0010_create_apps.sql
@@ -1,0 +1,8 @@
+CREATE TABLE `apps` (
+	`id` integer PRIMARY KEY NOT NULL,
+	`name` text NOT NULL,
+	`language` text NOT NULL,
+	`external_id` text NOT NULL,
+	`created_at` integer DEFAULT (unixepoch()) NOT NULL,
+	`updated_at` integer DEFAULT (unixepoch()) NOT NULL
+);