+++ packages/components/src/components/ui/switch.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as SwitchPrimitives from '@radix-ui/react-switch';
 
-import { cn } from '../../lib/utils';
+import { cn } from '../../lib/utils.js';
 
 const Switch = React.forwardRef<
   React.ElementRef<typeof SwitchPrimitives.Root>,